+++ src/reduxStore/selectors/batchTransactionsSelectors.ts
@@ -1,0 +1,10 @@
+import { RootState } from '../store';
+import { createDeepEqualSelector } from './helpers';
+
+export const batchTransactionsSelectors = (state: RootState) =>
+  state.batchTransactions;
+
+export const batchesSelector = createDeepEqualSelector(
+  batchTransactionsSelectors,
+  (state) => state
+);