+++ src/types/serverTransactions.types.ts
@@ -1,5 +1,6 @@
 import { AssetType, ScamInfoType } from './account.types';
 import { EsdtEnumType, NftEnumType } from './tokens.types';
+import { SignedTransactionType } from './transactions.types';
 
 //#region server trasactions
 export interface ScResultType {
@@ -309,3 +310,24 @@ export enum DecodeMethodEnum {
 }
 
 //#endregion
+
+export enum BatchTransactionStatus {
+  pending = 'pending',
+  success = 'success',
+  invalid = 'invalid',
+  dropped = 'dropped'
+}
+
+export interface BatchTransactionsResponseType {
+  id: string;
+  status: BatchTransactionStatus;
+  transactions: SignedTransactionType[] | SignedTransactionType[][];
+  error?: string;
+  message?: string;
+  statusCode?: string;
+}
+
+export type BatchTransactionsWSResponseType = {
+  batchId: string;
+  txHashes: string[];
+};